+++ src/main/java/com/andrewcomputsci/pathfinderfx/Model/Statistics.java
@@ -4,6 +4,6 @@ import com.andrewcomputsci.pathfinderfx.view.CellRectangle;
 
 import java.util.List;
 
-public record Statistics(List<CellRectangle> path, int passes, long deltaTime, int pathCost) {
+public record Statistics(List<CellRectangle> path, int passes, long deltaTime, double pathCost) {
 
 }
